+++ BE-MyCarMaster/src/main/java/softeer/bemycarmaster/api/domain/trim/dto/request/GetTrimsRequest.java
@@ -14,5 +14,5 @@ public class GetTrimsRequest {
 	@Schema(description = "모델 식별자", example = "1")
 	@NotNull(message = "modelId는 Null일 수 없습니다.")
 	@Min(value = 1, message = "modelId는 1 이상의 값입니다.")
-	private Integer modelId;
+	private Long modelId;
 }
